DnsNameResolver: Fail query if id space is exhausted (#13784)
Motivation:

When we try to execute a query we will try to select / generate an id
that is used. When we are not able to find one we throw an
IllegalStateException. In this case we also need to ensure we fail the
original promise as otherwise the user might never be notified of the
problem.

Modifications:

- Move throw code out of the DnsQueryContextManager to make it easier to
reason about
- Fail query with IllegalStateException

Result:

Query will be correctly failed in the case of id space exhausting