Updated HTTP2 Reader to fix missing header state
Motivation:
When a malformed request comes in where a stream error is thrown, data
is still left on the reader and not properly cleaned. When a valid
request comes in, it is not properly parsed because of the leftover data
being included.

Modifications:
Updated DefaultHttp2Reader to clean up leftover data if these types of
stream exceptions come in the wire.

Result:
After this change, data should be properly parsed after a malformed
request