DnsNameResolver: Add DnsQueryIdSpace class to reduce overhead while generating IDs
Motivation:

How we currently implement the selection of the next id to use for a query is very inefficient, especially once we already have a lot of inflight queries. This is problematic as we basically "block" the eventloop while doing the for loop.

Modifications:

Add new datastructure for storing and selecting ids that is basically O(1)

Result:

Less overhead while selecting ids