Backport #13075: Add the AdaptivePoolingAllocator
See PR #13075 for the motivations.

The AdaptivePoolingAllocator is modified to work with the ByteBuf API.
This required adding another implementation of ByteBuf, because in this API the buffer and their allocator (pooling or otherwise) are tightly coupled in their implementations.

Note that the AdaptivePoolingAllocator require at least Java 8, because it relies on StampedLock.
The constructor performs a version check, and throws an exception on older Java versions.