HTTP2: Update local window should not fail queued frames
Motivation:

When new stream frames arrive, they can be queued in
`AbstractHttp2StreamChannel.inboundBuffer` if the stream channel
doesn't have demand. When `stream.read()` is invoked, we first execute
`updateLocalWindowIfNeeded()`. For details, see #9400. However,
it may happen that the parent channel is already closed by that time.
As a result, write fails and an exception will be propagated through
the stream pipeline without allowing users to read frames that were
already received.

Modifications:

Taking into account the need to run `updateLocalWindowIfNeeded()` before
`doBeginRead()`, add a check for `updateLocalWindowIfNeeded()` to write
only if the parent context is not removed.

Result:

Users don't lose queued frames that were received before the parent
channel was closed.