Add constructors for UTF8 validation configuration to WebSocketClientProtocolHandler (#13813)
Motivation:

Ability to disable UTF8 validation was added in
https://github.com/netty/netty/commit/ebfc79c85deb8ba6fda54b8cb7fa925573076b21
but constructors in the WebSocketClientProtocolHandler were not updated
appropriately, preventing disabling of UTF8 validation when passing in
an externally-instantiated handshaker.

Modification:

This PR adds two additional constructors:
- A constructor that takes a WebSocketClientHandshaker and accepts a
withUTF8FrameValidation boolean as an addition to previously supported
config params.
- A constructor that takes a WebSocketClientHandshaker and a
WebSocketClientProtocolConfig as params, allowing for more convenient
instantiation and robustness to future WebSocketClientProtocolConfig
changes.

Result:

Allows instantiation of a WebSocketClientProtocolHandler which takes an
externally-provided handshaker and disables UTF8 validation.