DNS resolver final CNAME lookup disabled (#13850)
Follow up of https://github.com/netty/netty/pull/13721

Motivation:

The previous effort introduced a root-cause for NXDOMAIN and SERVFAIL.
That was only for the authoritative part. The non-authoritative was
slightly different because it allowed a CNAME lookup regardless, for
legacy reasons. In this effort, I take another stab at it, by making the
additional CNAME lookup an opt-in feature. Justification included in the
patch.

Modification:

Last resort CNAME lookup while doing address resolutions, is now an
opt-in feature by default. The non-authoritative flow also includes the
root cause as part of the UnknownHostException.

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>