Motivation: The DnsNameResolver does instantiate an instance FastThreadLocal for each instance of the resolver that is created. This can have an impact on the InternalThreadLocalMap static index an the size of allocated arrays for FastThreadLocal. In addition this FastThreadLocal is only used by the DnsNameResolver query methods and not the resolve methods. Common usage of the DnsNameResolver by Netty (i.e. by a Bootstrap) only uses resolves and queries are not performed.
Modification: This encapsulate the usage of the FastThreadLocal/dnsServerAddressStreamProvider in a specific implementation of DnsServerAddressStream that is instantiated by the DnsNameResolverBuilder by default to keep the current behavior. The builder now does provide a QueryServerAddressStream  property that can be set prior calling build in order to provide a specific implementation of the stream for use case that do not whish to sustain the cost of instantiating a non static FastThreadLocal.

Result: DnsNameResolver that do not use FastThreadLocal can now be created.

This provides a possible solution for #9884