Throw error on invalid cookie value Motivation:
Many browsers support cookie values separated by ';' instead of '; ', even if they violate the HTTP spec. HttpConversionUtil.toHttp2Headers would, however, silently drop a character of a second cookie in this case ('one=foo;two=bar' -> 'one=foo' and 'wo=bar').

Modification:

Check that the skipped character is actually a space.

Result:

An exception is thrown on invalid input, saving time for the next person that has to debug this :)