Fix x-www-form-urlencoded parsing for no-value key (#13904)
Motivation:

According to the specification for parsing of
application/x-www-form-urlencoded content at
https://url.spec.whatwg.org/#application/x-www-form-urlencoded, a key
without an = should be able to be parsed and given an empty value. The
current implementation of HttpPostStandardRequestDecoder fails to parse
these no-value keys when they are the last value in the sequence.

Modifications:

HttpPostStandardRequestDecoder is modified to include a key with no
value that is at the end of the undecoded chunk in the existing "special
empty FIELD" code path that currently only handles such fields when they
are followed by a '&' character.

Additional tests are provided to throroughly exercise variations of
content bodies with such empty fields.

Result:

Keys with no value that appear at the end of a x-www-form-urlencoded
sequence will be parsed according to the spec.