Remove the lock usage in PoolArena#numPinnedBytes() (#13906)
Motivation:

The implementation of `PoolArena#numPinnedBytes()` uses a lock to
compute the result of the method. This only purpose of this lock is to
provide an accurate computation of metric of the list of chunks: the
`chunkListMetrics` is final and an unmodifiable list. Note that the
returned value is already an estimate since the computation also sums
`activeBytesHuge` outside of the lock. In addition this method is used
by the allocator and is already an estimate as it sums the list of
arenas which can change during the iteration.

Modifications:

This removes the lock around the `chunkListMetrics` iteration in the
`PoolArena#numPinnedBytes()` method.

Result:

Less locking