Handle invalid cookie value during translation to http2 (#13891)
Motivation:

Many browsers support cookie values separated by ';' instead of '; ', even if they violate the HTTP spec. HttpConversionUtil.toHttp2Headers would, however, silently drop a character of a second cookie in this case ('one=foo;two=bar' -> 'one=foo' and 'wo=bar').

Modification:

Add a verification pass that checks that the cookie value can be split. If there is a semicolon without a space, or a character that `AsciiString` cannot handle, keep the cookie value as-is without splitting.

Result:

An invalid cookie is transmitted as-is instead.