Make /etc/resolv.conf reading more robust
Motivation:
It is possible that /etc/resolv.conf is missing or cannot be read, in which case `ResolvConf.system()` will throw an exception.
We need to handle that exception in `DefaultDnsServerAddressStreamProvider`, since exceptions thrown from a static initializer can become `NoClassDefFoundError`.

Modification:
Catch the exception that `ResolvConf.system()` might throw, an fall back to JNDI in that case.

Result:
No more `NoClassDefFoundError` being thrown when `/etc/resolv.conf` cannot be read for any reason.

Fixes #13949