Don't send a RST frame when closing the stream in a write future whil… (#13973)
…e processing inbound frames.

Motiviation:

Due a bug in netty we would send a RST frame in some cases even tho we
correctly received the endOfStream already. This is not necessary and
might even confuse the remote peer.

Modifications:

- Keep track of if we received endOfStream and send endOfStream in our
Channel implementation and only send a RST frame if this is not the case
during close
- Add unit tests

Result:

Don't send RST frame if we received endOfStream and send endOfStream.