Expose HTTP/2 decoder safety settings on Http2ConnectionHandlerBuilder (#13983)
Motivation:
We've added a number of safety settings for HTTP/2 decoding to the
AbstractHttp2ConnectionHandlerBuilder, and exposed them on
Http2FrameCodecBuilder and Http2MultiplexCodecBuilder, but not
Http2ConnectionHandlerBuilder.

Modification:
Add the missing overrides to Http2ConnectionHandlerBuilder so people who
use that API can configure these things as well.

Result:
It is now possible to configure the precise behavior of the mitigations
for Empty Frame Flooding (CVE-2019-9518) and Rapid Reset
(CVE-2023-44487) when using the Http2ConnectionHandlerBuilder.