Fix Zstd throws Exception on read-only volumes (#13978) (#13982)
Motivation:

When deploying Netty on a read-only volume (e.g, within a Docker
container) with "zstd-jni" included in the classpath, initialization of
zstd will fail because it requires unpacking files into a temporary
directory. Therefore, it is crucial to verify whether zstd can be
initialized, rather than merely checking for the presence of the library
in the classpath.

Modification:

During the verification process for the presence of "zstd-jni" in the
classpath, the system will also attempt to initialize the library. This
initialization follows a pattern similar to that used for the Brotli
library check.

Result:

Fixes #13978.

Co-authored-by: Daniel Käs <daniel.kaes@eviden.com>