Capture the AdaptiveByteBuf internal NIO buffer while holding magazine lock
We must never touch the chunk-delegate buffer's internal NIO buffer without holding the magazine lock.
Getting the internal NIO buffer involves modifying its position and limit, which are a shared mutable resource.
Therefor, we cannot lazily obtain this buffer, and have to grab it eagerly when the initialize the AdaptiveByteBuf under the magazine lock, which ensures that no other thread is touching that buffer.
This does unfortunately extend the critical section a bit.