ByteBuf Output Stream Reference Count Ownership (#13986)
Motivation:
Netty provides a adaptor from ByteBuf to Java's OutputStream interface.
The JDK Stream interfaces have an explicit lifetime because they
implement the Closable interface. This lifetime may be differnt than the
ByteBuf which is wrapped, and controlled by the interface which accepts
the JDK Stream. However Netty's ByteBufOutputStream currently does not
take reference count ownership of the underlying ByteBuf. There may be
no way for existing classes which only accept the OutputStream interface
to communicate when they are done with the stream, other than calling
close(). This means that when the stream is closed it may be appropriate
to release the underlying ByteBuf, as the ownership of the underlying
ByteBuf resource may be transferred to the Java Stream.

Modification:
Add the parameter releaseOnClose to ByteBufOutputStream.

Result:
ByteBufOutputStream can assume reference count ownership so the
underlying ByteBuf can be cleaned up when the stream is closed.

Related issue: #13985

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>