Add the AdaptivePoolingAllocator (#13075)
Motivation:
The current PoolingBufferAllocator has a number of issues. The code has
proven complex to maintain over the years, and has caused a steady
trickle of bugs despite its age. It also makes use of ThreadLocals,
which is problematic when used with VirtualThreads as there might be
millions of such threads, making the memory usage of
PoolThreadLocalCache prohibitively expensive.

Modification:
Instead of using ThreadLocals, the AdaptivePoolingAllocator reduces
contention by striping access across a number of magazines. Each thread
will have a preferred magazine to access, computed from their thread ID.
We use StampedLocks to tryLock the magazine, and if the lock is not
immediately available we will continue on to the next magazine. If there
is too much contention (too many tryLocks that fail) then the number of
magazines is increased, up to twice the number of CPU cores. This
behavior is similar in spirit to Stripe64, though the implementation is
quite different.

Each magazine does pointer-bump (using Buffer.split) allocations out of
a chunk. The default chunk size is 128 KiB.
However, allocation sizes are continuously recorded in a sliding-window
histogram. Every few thousand allocations to a magazine, the
99-percentile allocation size is computed from the histogram, and then
multiplied by 10. If this value is greater than 128 KiB, then it will be
used as the new preferred chunk size. From then on, chunks that are less
than this, or over 50% larger than this, will be retired and the memory
will be freed.

The maximum size we can track in the histograms is 1 MiB, and the
minimum size is 8192 bytes. The sizes are bucketed into powers of two.
This keeps our memory spent on histograms low. This also means our max
chunk size is 10 MiB.
We choose 8192 as the minimum bucket size because we also choose 128 KiB
as the minimum chunk size, so sizes of 8192 and smaller will have no
impact on the chunk size.
This in turn means we have 8 size buckets, and that allows the JIT to
optimize the loops over these arrays quite well.
The sliding window also only keep 4 generations of histograms, and we
use shorts for bucket counters since we rotate them every 1.024 to
65.534 allocations.
Because processing the histograms is relatively expensive, we adapt the
frequency of this processing depending on whether the computed chunk
size changed or not.

The allocator has a central queue where magazines can pluck available
chunks. However, magazines also have a nextInLine field, which will be
checked first - if it is empty we will instead take a chunk from the
central queue, or allocate a new one if the queue was empty. When all of
the buffers allocated from a chunk have been closed, then the chunk
returns to the allocator. When this happens we first check if the
nextInLine field of the originating magazine is null - if it is, we
atomically return the chunk to that magazine. Otherwise the chunk is
given to the central queue.

The dynamic adjustments we do to the chunk sizes and the number of
magazines, in response to the workload, is why this allocator is called
adaptive.

The adaptive allocator can be enabled as the Netty default allocator by
setting the `io.netty5.allocator.type` system property to `adaptive`.

Result:
We have a new pooling allocator with many benefits:

* It does not rely on thread locals.
* It uses less memory than the existing pooling allocator in the
practical applications that have been observed so far.
* It is inherently robust to fragmentation as long as buffers are
eventually closed.
* It is faster and (perhaps surprisingly) appear to scale better to
multi-threaded workloads than the existing allocator.