Reduce the likelihood of io_uring OP_CONNECT completions going to the wrong channel
Motivation:
File descriptor reuse can cause completions to be directed to the wrong channel.
This was a problem for connect() operations in our tests.
The problem manifests as spurious AlreadyConnectedExceptions being thrown from AbstractChannel.finishConnect().

Modification:
Submit a cancellation for any pending connect operation, when a channel is preparing to close.
Also attach a counter to the extra data for connect operations, and ignore OP_CONNECT completions where the counter stamp is not what we expected.

Result:
Although the counter is a 16-bit short, it still greatly reduces the likelihood of OP_CONNECT completions going to the wrong channel, since now both the counter and the file descriptor need to be reused in synchrony.