Update BouncyCastle to the latest version
Motivation:
The update was held back by a compatibility.
Since in 4.2 we've moved the JDK baseline from Java 6 to Java 8 anyway, we have the opportunity to bump the BC version as well.

Modification:
Switch all BouncyCastle dependencies from "-jdk15on" to "-jdk18on", and update them to the latest version.

Result:
Fixes #13561