EmbeddedChannel: Ensure tasks are not executed while we already execute something else
Motivation:

We need to ensure we only try to run tasks that were submitted for exectuion when we are not currently run other code in the EmbeddedChannel. We failed to do so and so coudl end up executing tasks directly while these should have been only run once the current execution is complete.

Modifications:

- Keep track of how deep our "execution stack" is and only run the pending tasks if we are sure it will not cause out of order problems
- Testcase

Result:

Correctly respect ordering