Keep used NioIoOps and SelectionKey in sync
Motivation:

As the user is able to retieve the underlying SelectionKey via the NioRegistration we need to ensure we always keep this and the current used NioIoOps in sync.

Modifications:

Remove the local stored NioOps and just always reconstruct using the underlying SelectionKey. This is not a concern in terms of performance and object creation as we pre-construct them.

Result:

Keep current used NioIoOps and SelectionKey always in sync