Remove some reflection usage
Motivation:
We've been using reflection as a way to abstract over Java API versions.
Now that we're based on Java 8, some of this reflection usage is no longer necessary.
In some cases, the APIs we were reflecting on are now directly available.
In other cases, we'd like a little more performance and can call through method handles instead.

Modification:
Remove some reflection usage that was necessary when running on Java 6 or 7, and replace it with direct calls if the code is still needed.
Replace the more performance sensitive reflection usage with MethodHandles.

Result:
Cleaner, and perhaps slightly faster, code.