Optimize wrap buffer cumulation in SslHandler and don't mutate input buffers  (#14086)
Motivation:

SslHandler doesn't support sharing the input buffers since the input
buffers get mutated. This issue has been reported as #14069. Fixes have
already been made to support the use case using read-only buffers as
input. However, this is not useful for existing Netty applications.

I checked the Netty code once again and noticed that it's possible to
optimize the current implementation in SslHandler and avoid the buffer
mutation.

The default behavior in SslHandler causes very hard to debug problems.
In Pulsar, there have been issues open for multiple years that have been
caused by the SslHandler's default behavior of mutating the input
buffers (apache/pulsar#22601 is one starting point to the issues).

The current implementation of SslHandler will mutate the input buffers
by using the input buffer as a cumulation buffer when there's available
capacity. This is not optimal at all. This PR contains an improvement
where the cumulation buffer is allocated to the final size immediately
so that cumulation is more optimal. That's why I think that this will be
more optimal in the end and won't cause performance regressions.

After these changes, it should be possible to pass a slice or duplicate
of a buffer and the input buffer doesn't get mutated. The input buffer's
readIndex will be modified so multiple threads cannot share the same
input buffer without a slice or duplicate.

I created a failing test case that shows that a duplicate buffer doesn't
prevent input buffer mutation.

![image](https://github.com/netty/netty/assets/66864/0abec333-6079-40bd-8bc1-ccfd3df6fae1)
It's a very surprising behavior of SslHandler that passing a
`.retainedDuplicate()` buffer could result in mutations to the original
buffer. That's another reason why I think that SslHandler shouldn't ever
mutate the inputs.

Modification:

- add `protected ByteBuf composeFirst(ByteBufAllocator allocator,
ByteBuf first, int bufferSize)` to AbstractCoalescingBufferQueue to
support allocating an optimal cumulation buffer instead of using the
input buffer as a cumulation buffer
- remove dead code from `compose` where there was handling for
CompositeByteBuf
- `composeFirst` will be called before `compose` so this case isn't
needed.

Result:

Fixes #14069 without the need to set the input buffer as read-only

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>