Fix AdaptiveByteBufAllocator class loading on Java 6/7
Motivation:
We cannot have any class that *mention* Java 8+ APIs like StampedLock, before we've passed a version check.
The AdaptiveByteBufAllocator referenced the AdaptivePoolingAllocator via fields, which in turn referenced StampedLock via fields, thus AdaptiveByteBufAllocator cannot be loaded on Java 6/7.

Modification:
Hide AdaptivePoolingAllocator behind an interface in the AdaptiveByteBufAllocator, and move the version check to the AdaptiveByteBufAllocator constructor before we build the inner allocators.

Result:
The AdaptiveByteBufAllocator class can now hopefully load on Java 6/7, and get to the point where we throw the exception from the constructor.