Rethrow Error and RuntimeException when caused by invokeExact(...)
Motivation:

We should just rethrow the Error / RuntimeException when caused by invokeExact(...) as always wrapping it in a LinkageError might be confusing. For example we might encounter an OutOfMemoryException or an InvalidArgumentException.

Modifications:

Check if we should just rethrow or not

Result:

Less confusing and more correct handling of exceptions causes by invokeExact(...)