Configure hostname validation via SslContextBuilder
Motivation:
Hostname validation is an important step for clients, in verifying the identity of the server.

We should make it easy to configure.

Due to its importance, we should also enable it by default.
While there are theoretically multiple algorithms to choose from for this purpose, the "HTTPS" algorithm is *by far* the most commonly used one.

Modification:
- Add a SslContextBuilder.endpointIdentificationAlgorithm() method to configure hostname validation.
- Set the endpoint validation algorithm to "HTTPS" by default, when creating SslContextBuilders for clients.

Result:
Hostname validation is now enabled by default for clients, and can easily be configured through the SslContext.