Fix the graal build
Motivation:
In https://github.com/netty/netty/pull/14032 we configured Graal to prepare reflective access to all public methods on the jdk.internal.misc.Unsafe.
This apparently includes linking to the getUncompressedObject method, which I think only exist in HotSpot and not in Graal, so the build fails with a linkage error.

Modification:
Only prepare reflective access for the methods we'll actually call.

Result:
No more Graal linkage errors during the build.