Null value doesn't disable endpoint verification for JDK provider
Motivation:

If Java ever changes the default endpoint identification algorithm
of if users use a 3-party provider that overrides the default,
Netty users won't be able to change it via `SslContextBuilder` API by
providing a null value. Only an empty string will disable it, but this
behavior is not documented in `endpointIdentificationAlgorithm(String)`
javadoc.

Modifications:
- Set `endpointIdentificationAlgorithm` unconditionally in
`JdkSslContext`, similar to `ReferenceCountedOpenSslServerContext`.

Result:

`SslContextBuilder.endpointIdentificationAlgorithm(null)` works as
expected with `SslProvider.JDK`.