Add support to suspend EventExecutor
Motivation:

At the moment once an EventExecutor is created its impossible to suspend it, which means it will stay alive until shutdown. While this is generally fine it also might result in more system usage then really necessary.
For example an EventExecutor might tie up an Thread for its whole life-time where it would be possible to "release" the Thread back one it was suspended.

Modifications:

- Allow to suspend EventExecutors (if supported by the implementation
- Add unit tests

Result:

Allow to suspend EventExecutor (and subtypes)