fix ResolvConf initialization with SecurityManager enabled
Motivation:

JNDI fallback in DefaultDnsServerAddressStreamProvider will not work,
if SecurityManager blocks access to /etc/resolv.conf.
Handle Exception for correct ResolvConf initialization.

The ResolvConf lookup was added in #13884 and the fallback added in #13951

Modification:

Catch a SecurityException and keep the resolvConf null.

Result:

Initialization will work with restrictive SecurityManager config.