Backport the SslContextBuilder.endpointIdentificationAlgorithm method
Motivation:
We added this method in 4.2, and set the default endpoint identification algorithm to "HTTPS".
To keep the APIs as similar as possible, and to allow integrators to start referencing the new method early, we backport it to 4.1.
The default endpoint identification algorithm default remains unchanged in 4.1, where it is disabled as it always has been.
At least now it's easier to enable.

Modification:
- Backport the API changes from #14127 (8372a09092e84b98e55facfe3bd4c23448c32918) and #14144 (b9552ea9d81427252303352ad7005b8651ea52b7).
- Keep the default setting the same, so no tests needs to change.
- Guard the actual SSLParameters API call by a Java 7 version check, which is now necessary since the compatibility baseline is Java 6.

Result:
Netty 4.2 and 4.1 now share the same API for enabling hostname verification.
It's easier to enable hostname verification.