Don't strip whitespaces from header names and let the validator handle these
Motivation:

We did strip whitespaces from the header names before trying to add these to DefaultHeaderName. We should better not do this as whitespaces as prefix of a header name are not valid in general.

Modifications:

Don't strip leading whitespaces and so let the validator handle it.

Result:

Fixes https://github.com/netty/netty/issues/14177