Add keytool-based self-signed certificate generator (#14214)
Forward port of #14198

Motivation:

On newer Java versions, the JDK-based self-signed certificate generator
does not work, so adding bcpkix as a dependency was necessary. This is
inconvenient for users that want to use self-signed certs as part of
their tests.

Modification:

Add a KeytoolSelfSignedCertGenerator which uses the keytool command
included in the JDK to generate the key pair and certificate.

Introduce a SelfSignedCertificate.Builder to make the constructor chaos
a bit cleaner. Additionally, since the keytool generator uses an
external RNG, this Builder allows for the in-vm KeyPair generation to
happen lazily, improving performance for the keytool path.

The next step I've thought about doing is to move the unnecessary file
system saving (SelfSignedCertificate.newSelfSignedCertificate) to be
lazy, so that we can avoid writing the key and cert at all in some
cases. But that is for another day, it'd require some exception handling
changes.

Result:

Users can generate self-signed certificates without additional
dependencies.