Revert usage of MethodHandles for sun.misc.Unsafe (#14217)
Motivation:

acfebeb introduced the usage of MethodHandles for sun.misc.Unsafe access
to be able to compile with --release 8. While this works it also means
that there is overhead to lookup things during initialization which
negativly effects startup times. This is problematic for things like
quakus and others.

Modifications:

- Partly revert commit acfebeb65865ae05b26a2d08d5f90f8385862cb8.
- Just depend on stubs for sun.misc.Unsafe with provided scope. This way
we can still use --release 8 during compilation and not add the extra
overhead

Result:

Reduce startup time of apps