Ensure AbstractCoalescingBufferQueue does not end up in inconsistent state on error
Motivation:

The code in AbstractCoalescingBufferQueue.remove(...) could lead to inconsistent internal state when for example an exception is thrown because of reference counting problems. Because of this is was possible that while the queue was empty it still showed readableBytes() > 0. This could then lead to various issues internally within netty where we depend on that the queue will eventually returns readableBytes() == 0.

Modifications:

- Ensure we always decrement readable bytes, even in the cause of an error
- Also correctly remove the listener if an error happens during handling the buffer.

Result:

Fixes https://github.com/netty/netty/issues/11959