Removal of split package between netty-codec and netty-codec-xml
Motivation:

The netty-codec maven module contains the io.netty.handler.codec.xml.XmlFrameDecoder class which shares a package with the package of the netty-codec-xml maven module, and therefore exhibit a split package. A split package is not allowed by the java module system.

Changes:

The class io.netty.handler.codec.xml.XmlFrameDecoder is moved from the netty-codec maven module to the netty-codec-xml maven module along with its tests. Since it is a breaking change the revapi difference has been added to the list of ignored breaking changes.

Results:

The split package is removed, users of the class io.netty.handler.codec.xml.XmlFrameDecoder will need to add the netty-codec-xml maven module to their dependencies in order to continue using the class.